+++ packages/core/solidity/src/erc20.ts
@@ -13,6 +13,7 @@ import type { ClockMode } from './set-clock-mode';
 import { clockModeDefault, setClockMode } from './set-clock-mode';
 import { supportsInterface } from './common-functions';
 import { OptionsError } from './error';
+import { toUint256, UINT256_MAX } from './utils/convert-strings';
 
 export const crossChainBridgingOptions = [false, 'custom', 'superchain'] as const;
 export type CrossChainBridging = (typeof crossChainBridgingOptions)[number];
@@ -163,24 +164,6 @@ export function isValidChainId(str: string): boolean {
   return chainIdPattern.test(str);
 }
 
-const UINT256_MAX = BigInt(2) ** BigInt(256) - BigInt(1);
-
-function toUint256(value: string, field: string): bigint {
-  const isValidNumber = /^\d+$/.test(value);
-  if (!isValidNumber) {
-    throw new OptionsError({
-      [field]: 'Not a valid number',
-    });
-  }
-  const numValue = BigInt(value);
-  if (numValue > UINT256_MAX) {
-    throw new OptionsError({
-      [field]: 'Value is greater than uint256 max value',
-    });
-  }
-  return numValue;
-}
-
 function scaleByPowerOfTen(base: bigint, exponent: number): bigint {
   if (exponent < 0) {
     return base / BigInt(10) ** BigInt(-exponent);
@@ -207,18 +190,8 @@ function addPremint(
       const units = integer + decimals + zeroes;
       const exp = decimalPlace <= 0 ? 'decimals()' : `(decimals() - ${decimalPlace})`;
 
-      // Validate base units is within uint256 range
       const validatedBaseUnits = toUint256(units, 'premint');
-
-      // Check for potential overflow assuming decimals() = 18
-      const assumedExp = decimalPlace <= 0 ? 18 : 18 - decimalPlace;
-      const calculatedValue = scaleByPowerOfTen(validatedBaseUnits, assumedExp);
-
-      if (calculatedValue > UINT256_MAX) {
-        throw new OptionsError({
-          premint: 'Amount would overflow uint256 after applying decimals',
-        });
-      }
+      checkPotentialPremintOverflow(validatedBaseUnits, decimalPlace);
 
       c.addConstructorArgument({ type: 'address', name: 'recipient' });
 
@@ -247,6 +220,24 @@ function addPremint(
   } else {
     throw new OptionsError({
       premint: 'Not a valid number',
+    });
+  }
+}
+
+/**
+ * Check for potential premint overflow assuming the user's contract has decimals() = 18
+ *
+ * @param baseUnits The base units of the token, before applying power of 10
+ * @param decimalPlace If positive, the number of assumed decimal places in the least significant digits of `validatedBaseUnits`. Ignored if <= 0.
+ * @throws OptionsError if the calculated value would overflow uint256
+ */
+function checkPotentialPremintOverflow(baseUnits: bigint, decimalPlace: number) {
+  const assumedExp = decimalPlace <= 0 ? 18 : 18 - decimalPlace;
+  const calculatedValue = scaleByPowerOfTen(baseUnits, assumedExp);
+
+  if (calculatedValue > UINT256_MAX) {
+    throw new OptionsError({
+      premint: 'Amount would overflow uint256 after applying decimals',
     });
   }
 }
